StackOverFlowApi/backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework==3.14.0
idna==3.4
pycodestyle==2.10.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
